Data.HashMap.Monoidal

Description

This module provides a HashMap variant which uses the value's Monoid instance to accumulate conflicting entries when merging Maps.

While some functions mirroring those of HashMap are provided here for convenience, more specialized needs will likely want to use either the Newtype or Wrapped instances to manipulate the underlying Map.

Documentation

newtype MonoidalHashMap k a

A HashMap with monoidal accumulation

Constructors

MonoidalHashMap

Fields

getMonoidalHashMap :: HashMap k a

Often-needed functions

toList :: MonoidalHashMap k a -> [(k, a)]

O(n*log n). Return a list of this map's elements. The list is produced lazily. The order of its elements is unspecified.

fromList :: (Eq k, Hashable k) => [(k, a)] -> MonoidalHashMap k a

O(n*log n). Construct a map with the supplied mappings. If the list contains duplicate mappings, values will be replaced.

fromListWith :: (Eq k, Hashable k) => (a -> a -> a) -> [(k, a)] -> MonoidalHashMap k a

O(n*log n). Construct a map with the supplied mappings. If the list contains duplicate mappings, values will be merged using the provided combining function.

singleton :: (Eq k, Hashable k) => k -> a -> MonoidalHashMap k a

O(1). A map with a single element.

size :: MonoidalHashMap k a -> Int

O(1). The number of elements in the map.

member :: (Eq k, Hashable k) => k -> MonoidalHashMap k a -> Bool

O(log n). Is the key a member of the map? See also notMember.

notMember :: (Eq k, Hashable k) => k -> MonoidalHashMap k a -> Bool

O(log n). Is the key not a member of the map? See also member.

lookup :: (Eq k, Hashable k) => k -> MonoidalHashMap k v -> Maybe v

O(log n) Return the value to which the specified key is mapped, or Nothing if this map contains no mapping for the key.

lookupM :: (Eq k, Hashable k, Monoid v) => k -> MonoidalHashMap k v -> v

O(log n) Return the value to which the specified key is mapped, or mempty if this map contains no mapping for the key.

elems :: MonoidalHashMap k a -> [a]

O(n). Return a list of this map's values. The list is produced lazily.

keys :: MonoidalHashMap k a -> [k]

O(n). Return all keys of the map in ascending order. Subject to list fusion.

delete :: (Eq k, Hashable k) => k -> MonoidalHashMap k a -> MonoidalHashMap k a

O(log n). Delete a key and its value from the map. When the key is not a member of the map, the original map is returned.

mapKeys :: (Hashable k, Eq k, Hashable k', Eq k') => (k -> k') -> MonoidalHashMap k a -> MonoidalHashMap k' a

O(n). Map a function to each key of a map, if it will result in duplicated mappings, their values will be merged in unspecified order

insert :: (Hashable k, Eq k) => k -> a -> MonoidalHashMap k a -> MonoidalHashMap k a

O(log n). Insert a value on some key, if it exists replace the value.

insertWith :: (Hashable k, Eq k) => (a -> a -> a) -> k -> a -> MonoidalHashMap k a -> MonoidalHashMap k a

O(log n). Insert a value on some key, if it exists apply the combining function.

modify :: (Monoid a, Hashable k, Eq k) => (a -> a) -> k -> MonoidalHashMap k a -> MonoidalHashMap k a

O(log n). Modify a value on some key with a function, if value under key doesn't exist -- use mempty.

modifyDef :: (Hashable k, Eq k) => a -> (a -> a) -> k -> MonoidalHashMap k a -> MonoidalHashMap k a

O(log n). Modify a value on some key with a function, providing a default value if that key doesn't exist.

map :: (v1 -> v2) -> MonoidalHashMap k v1 -> MonoidalHashMap k v2

O(n) Transform this map by applying a function to every value.

filterWithKey :: (k -> v -> Bool) -> MonoidalHashMap k v -> MonoidalHashMap k v

O(n) Filter this map by retaining only elements satisfying a predicate.